Mouse: Change mouse coordinates to Low Battery OK button
Modify the coordinates such that instead of clicking the center of the
screen, we click the OK button on the low battery warning dialog.

Rename the method according to its new intent.

TODO: This probably doesn't work on other displays that have different
resolutions. We need to check that and figure out a cross-display
solution.