Remove MainMenu.xib
Since this is a backround app
(3db243f3c84f7ab36747f4badfd4f57a26563821), we don't have a UI and have
no need for any nib files.

We'll make an additional target as a separate application that does
have a UI and will allow customisation of the global hotkey.