Configure MASShortcut
Make some changes to the project to get MASShortcut working. Can't
remember what they were, it was yesterday.

Change the MASShortcut submodule URL to my fork of the project, which
includes changes to the Xcode project file and codebase that allow it to
compile in this project and on my development machine.