Use DDHotKeyTextField in "Low Battery Yup"
Add a text field to the MainMenu.xib primary window and use it to allow
users to type hotkeys. I'm thinking we might use this instead of
MASShortcut since I've been having trouble getting that to work
(couldn't figure out how to get a registered hotkey out of
NSUserDefaults).

This tool seems simpler, and could be just the ticket.