Able to get a shortcut from the MASShortcutView
Fuck damn it finally freaking works! Damn, just had to take a 1.5 hour
nap after that.

We're finally able to extract a shortcut out of the MASShortcutView.

After trying a whole bunch of different things, I ultimately tried
opening the debugger and figured out that for whatever reason the
bindings weren't getting created and because of that my shortcut wasn't
getting saved in the user defaults.

Couldn't figure out how to get the binding to work even after includes
and whatnot.

Finally decided screw it I'm going to subclass the view and get the
shortcut and store it in `NSUserDefaults` myself. So that's what we're
going to do and this is the first step to getting that working.

Hoo boy.

Definitely need to vaccum the code after this.