MainMenu.xib: Fix reference to `_shortcut_view` ivar
Broke the reference to `_shortcut_view` in
a984e2dd951bcbfe5e3e6374213eb9a5fd5423cf so when we called methods on
it, they didn't work. This fixes it. Oh man that gave me a hard time.
Glad I took a break and had an idea to call methods on the superclass
and see if they worked. They did, but not methods on `ShorcutView`. So
from that I figured out that I needed to update this reference.