Add .gitattributes for merging Xcode *.pbxproj
Automatically merge this file. I got conflicts in it trying to do a
merge and it seemed like a bad time. Not conflicts I should be resolving
myself manually.

Got the tip from a Stack Overflow answer by Josh Brown:
http://stackoverflow.com/questions/31541624/merge-conflicts-with-project-pbxproj-on-git-xcode-ios/32567843#32567843

According to this, though:
http://stackoverflow.com/questions/12907605/xcode-project-file-git-merge-conflict
that won't work in all cases. We'll try it anyway and hope for the best.