MainMenu.xib: Bind "Start at login" checkbox to NSUserDefaultsController
Add a Key-Value Binding in Interface Builder to
`NSUserDefaultsController` so that the checkbox value gets persisted
between launches of the app.

This fixes the bug mentioned in a77a053f90a7b2588dbaabedbf25f908680f2f30
such that now unchecking the box and closing the app will result in the
checkbox being unchecked on the next launch.

Wow, that was super easy!