Launch the daemon when "Start at login" is checked
Launch the daemon when checked and quit it when unchecked.

Now that I'm committing this, I don't really think this is the exact
right behaviour. I think running and terminating the app should be maybe
a different control so that running it isn't contingent on having it
start at login.