Add a button to launch & quit the daemon
Add a new button to the UI that either launches or quits the daemon
depending on whether it's currently running.

The button is set as a "Toggle" type that will cycle between "Launch
Application" and "Quit Application" labels when clicked.

Depending on the button's state (0 or 1), it will either launch or quit
the daemon.

The button's initial state is set on the UI application's launch to
determine how it should read & what it should do.