Copy MASShortcut.framework to bundle Frameworks/ directory
The framework was getting copied to Resources/, but the app was told to
look for it in Frameworks/, so it was crashing on launch.

Here's the error:

    Dyld Error Message:
      Library not loaded: @executable_path/../Frameworks/MASShortcut.framework/Versions/A/MASShortcut
      Referenced from: /Users/USER/Documents/*/Low Battery Yup.app/Contents/MacOS/Low Battery Yup
      Reason: image not found

Thank goodness for Console and logging and crash logs. Never would have
figured this out otherwise. Worked fine in Debug configuration and when
building for Release in Xcode. Just didn't work in the final archived
app. Works now though, phew!